Create Rating-Algorithim.jar
This is the algorithm we wrote to give each house a rating out of 100. 50/100 represents the average home.  All of our ratings are relative to Toronto averages. For example, the average house in Toronto costs $1000 per square foot, and so the average listing price would be 1 million dollars. Assuming it had the average number of bedroom (2) and bathroom (2), and no included amenities (pool, extra parking space, garage etc), it would receive a score of 50/100.  The lower the price per square foot the higher the rating, and the higher the included amenities the higher the rating.